Transient.Backtrack

Description

https://www.fpcomplete.com/user/agocorona/the-hardworking-programmer-ii-practical-backtracking-to-undo-actions

Documentation

registerUndo :: TransientIO a -> TransientIO a

register an action that will be executed when backtracking

onUndo :: TransientIO a -> TransientIO a -> TransientIO a

the secod parameter will be executed when backtracking

undo :: TransientIO a

execute backtracking. It execute the registered actions in reverse order.

If the backtracking flag is changed the flow proceed forward from that point on.

If the backtrack stack is finished or undoCut executed, undo will stop.

retry :: TransIO ()

restart the flow forward from this point on

undoCut :: TransientIO ()

assures that backtracking will not go further back