Transient.Internals

Description

See http://github.com/agocorona/transient everithing in this module is exported in order to allow extensibility.

Documentation

data TransIO x

Constructors

Transient

Fields

runTrans :: StateT EventF IO (Maybe x)

type SData = ()

type EventId = Int

type TransientIO = TransIO

data EventF

Constructors

EventF

Fields

meffects :: Effects

event :: Maybe SData

xcomp :: TransIO a

fcomp :: [b -> TransIO b]

mfData :: Map TypeRep SData

mfSequence :: Int

threadId :: ThreadId

freeTh :: Bool

parent :: Maybe EventF

children :: TVar [EventF]

maxThread :: Maybe (IORef Int)

type Effects = forall a b c. TransIO a -> TransIO a -> (a -> TransIO b) -> StateIO (StateIO (Maybe c) -> StateIO (Maybe c), Maybe a)

type StateIO = StateT EventF IO

runTransient :: TransIO x -> IO (Maybe x, EventF)

run the transient computation with a blank state

runTransState :: EventF -> TransIO x -> IO (Maybe x, EventF)

run the transient computation with an state

getCont :: TransIO EventF

get the continuation context: closure, continuation, state, child threads etc

runCont :: EventF -> StateIO (Maybe a)

run the closure and the continuation using the state data of the calling thread

runCont' :: EventF -> IO (Maybe a, EventF)

run the closure and the continuation using his own state data

getContinuations :: StateIO [a -> TransIO b]

warning: radiactive untyped stuff. handle with care

compose :: (Alternative f, Monad f) => [t -> f t] -> t -> f a

compose a list of continuations

runClosure :: EventF -> StateIO (Maybe a)

run the closure (the x in 'x >>= f') of the current bind operation.

runContinuation :: EventF -> a -> StateIO (Maybe b)

run the continuation (the f in 'x >>= f') of the current bind operation

setContinuation :: TransIO a -> (a -> TransIO b) -> [c -> TransIO c] -> StateIO ()

runContinuations :: [a -> TransIO b] -> c -> TransIO d

run a chain of continuations. It is up to the programmer to assure by construction that each continuation type-check with the next, that the parameter type match the input of the first continuation. Normally this makes sense if it stop the current flow with stop after the invocation

newtype IDNUM

Constructors

IDNUM Int

restoreStack :: MonadState EventF m => [b -> TransIO b] -> m ()

readWithErr :: (Typeable * a, Read a) => [Char] -> IO [(a, String)]

readsPrec' :: (Typeable * a, Read a) => t -> [Char] -> [(a, String)]

data IDynamic

dynamic serializable data for logging

Constructors

IDyns String

(Read a, Show a, Typeable a) => IDynamic a

type Recover = Bool

type CurrentPointer = [LogElem]

type LogEntries = [LogElem]

data LogElem

Constructors

Wait

Exec

Var IDynamic

data Log

Constructors

Log Recover CurrentPointer LogEntries

data RemoteStatus

Constructors

WasRemote

WasParallel

NoRemote

stop :: Alternative m => m stop

a sinonym of empty that can be used in a monadic expression. it stop the computation and execute the next alternative computation (composed with <|>)

(**>) :: TransIO a -> TransIO b -> TransIO b infixr 1

executes the second operand even if the frist return empty. A normal imperative (monadic) sequence uses the operator (>>) which in the Transient monad does not

(<***) :: TransIO a -> TransIO b -> TransIO a infixr 1

forces the execution of the second operand even if the first stop. Return the first result. The second operand is executed also when internal events happens in the first operand and it returns something

atEnd :: TransIO a -> TransIO b -> TransIO a

(<**) :: TransIO a -> TransIO b -> TransIO a infixr 1

forces the execution of the second operand even if the first stop. It does not execute the second operand as result of internal events occuring in the first operand. Return the first result

atEnd' :: TransIO a -> TransIO b -> TransIO a

(<|) :: TransIO a -> TransIO b -> TransIO a

when the first operand is an asynchronous operation, the second operand is executed once (one single time) when the first completes his first asyncronous operation.

This is useful for spawning asynchronous or distributed tasks that are singletons and that should start when the first one is set up.

for example a streaming where the event receivers are acivated before the senders.

setEventCont :: TransIO a -> (a -> TransIO b) -> StateIO EventF

set the current closure and continuation for the current statement

resetEventCont :: MonadState EventF m => Maybe t1 -> t -> m (a -> a)

reset the closure and continuation. remove inner binds than the previous computations may have stacked in the list of continuations. resetEventCont :: Maybe a -> EventF -> StateIO (TransIO b -> TransIO b)

tailsafe :: [t] -> [t]

baseEffects :: Effects

Threads

waitQSemB :: (Ord a, Num a) => IORef a -> IO Bool

signalQSemB :: Num a => IORef a -> IO ()

threads :: Int -> TransIO a -> TransIO a

set the maximun number of threads for a procedure. It is useful to limit the parallelization of transient code that uses parallel spawn and waitEvents

oneThread :: TransIO a -> TransientIO a

delete all the previous childs generated by the expression taken as parameter and continue execution of the current thread.

showThreads :: TransIO empty

addThreads' :: Int -> TransIO ()

add n threads to the limit of threads. If there is no limit, it set it

addThreads :: Int -> TransIO ()

assure that at least there are n threads available

freeThreads :: TransIO a -> TransIO a

The threads generated in the process passed as parameter will not be killed by `kill*` primitives

hookedThreads :: TransIO a -> TransIO a

The threads will be killed when the parent thread dies. That is the default. This can be invoked to revert the effect of freeThreads

killChilds :: TransientIO ()

kill all the child threads of the current thread

extensible state: session data management

getData :: (MonadState EventF m, Typeable a) => m (Maybe a)

Get the state data for the desired type if there is any.

getSData :: Typeable a => TransIO a

getData specialized for the Transient monad. if Nothing, the monadic computation does not continue.

If there is no such data, getSData silently stop the computation. That may or may not be the desired behaviour. To make sure that this does not get unnoticed, use this construction:

 getSData <|> error "no data"

To have the same semantics and guarantees than get, use a default value:

getInt= getSData <|> return (0 :: Int)

The default value (0 in this case) has the same role than the initial value in a state monad. The difference is that you can define as many get as you need for all your data types.

To distingish two data with the same types, use newtype definitions.

setData :: (MonadState EventF m, Typeable a) => a -> m ()

set session data for this type. retrieved with getData or getSData Note that this is data in a state monad, that means that the update only affect downstream in the monad execution. it is not a global state neither a per user or per thread state it is a monadic state like the one of a state monad.

delData :: (MonadState EventF m, Typeable a) => a -> m ()

genId :: MonadState EventF m => m Int

generator of identifiers that are unique withing the current monadic sequence They are not unique in the whole program.

getPrevId :: MonadState EventF m => m Int

data StreamData a

async calls

Constructors

SMore a

SLast a

SDone

SError SomeException

waitEvents :: IO b -> TransIO b

variant of parallel that repeatedly executes the IO computation and kill the previously created childs

It is useful in single threaded problems where each event discard the computations spawned by previous events

waitEvents' :: IO b -> TransIO b

async :: IO b -> TransIO b

variant of parallel that execute the IO computation once, and kill the previous child threads

spawn :: IO b -> TransIO b

variant that spawn free threads. Since there is no thread control, this is faster

parallel :: IO (StreamData b) -> TransIO (StreamData b)

return empty to the current thread, in new thread, execute the IO action, this IO action modify an internal buffer. then, executes the closure where parallel is located In this new execution, since the buffer is filled, parallel return the content of this buffer. Then it launch the continuation after it with this new value returned by the closure.

If the maximum number of threads, set with threads has been reached parallel perform the work sequentially, in the current thread. So parallel means that 'it can be parallelized if there are thread available'

if there is a limitation of threads, when a thread finish, the counter of threads available is increased so another parallel can make use of it.

The behaviour of parallel depend on StreamData; If SMore, parallel will excute again the IO action. with SLast, SDone and SError, parallel will not repeat the IO action anymore.

loop :: EventF -> IO (StreamData t) -> IO ()

forkFinally1 :: IO a -> (Either SomeException a -> IO ()) -> IO ThreadId

free :: ThreadId -> EventF -> IO (Maybe EventF)

hangThread :: EventF -> EventF -> IO ()

killChildren :: TVar [EventF] -> IO ()

kill all the child threads associated with the continuation context

type EventSetter eventdata response = (eventdata -> IO response) -> IO ()

type ToReturn response = IO response

react :: Typeable eventdata => EventSetter eventdata response -> ToReturn response -> TransIO eventdata

deinvert an event handler.

The first parameter is the setter of the event handler to be deinverted. Usually it is the primitive provided by a framework to set an event handler

the second parameter is the value to return to the event handler. Usually it is `return()`

it configures the event handler by calling the setter of the event handler with the current continuation

non-blocking keyboard input

getLineRef :: TVar (Maybe a)

roption :: MVar [t]

option :: (Typeable b, Show b, Read b, Eq b) => b -> String -> TransIO b

install a event receiver that wait for a string and trigger the continuation when this string arrives.

input :: (Typeable a, Read a, Show a) => (a -> Bool) -> String -> TransIO a

validates an input entered in the keyboard in non blocking mode. non blocking means that the user can enter also anything else to activate other option unlike option, wich watch continuously, input only wait for one valid response

getLine' :: (Typeable * a, Read a) => (a -> Bool) -> IO a

non blocking getLine with a validator

reads1 :: (Typeable * a, Read a) => [Char] -> [(a, String)]

inputLoop :: IO b

processLine :: String -> IO ()

rexit :: MVar a

stay :: IO b

wait for the execution of exit and return the result

keep :: TransIO a -> IO a

keep the main thread running, initiate the non blocking keyboard input and execute the transient computation.

It also read a slash-separated list of string that are read by option and input as if they were entered by the keyboard

 foo  -p  options/to/be/read/by/option/and/input

keep' :: TransIO a -> IO a

same than keepbut do not initiate the asynchronous keyboard input. Useful for debugging

exit :: a -> TransIO a

force the finalization of the main thread and thus, all the Transient block (and the application if there is no more code)

exit' :: MonadIO m => a -> m ()

onNothing :: Monad m => m (Maybe b) -> m b -> m b

alternative operator for maybe values. Used in infix mode

Orphan instances

Read SomeException

Methods

readsPrec :: Int -> ReadS SomeException

readList :: ReadS [SomeException]

readPrec :: ReadPrec SomeException

readListPrec :: ReadPrec [SomeException]